+++ storage/perfschema/table_threads.cc
@@ -187,7 +187,7 @@ int table_threads::read_row_values(TABLE *table,
 {
   Field *f;
   const char *str= NULL;
-  int len= 0;
+  size_t len= 0;
 
   if (unlikely(! m_row_exists))
     return HA_ERR_RECORD_DELETED;
@@ -298,7 +298,7 @@ int table_threads::read_row_values(TABLE *table,
       case 15: /* CONNECTION_TYPE */
         str= vio_type_name(m_row.m_connection_type, & len);
         if (len > 0)
-          set_field_varchar_utf8(f, str, len);
+          set_field_varchar_utf8(f, str, (uint)len);
         else
           f->set_null();
         break;
